requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asgi-lifespan==2.1.0
asyncpg==0.28.0
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==1.1.2
h11==0.13.0
httpcore==0.18.0
httpx==0.25.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.2.1
requests==2.28.1
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.19.0
